face detection/leo_face/requirements.txt
# Leo Face Recognition System – Dependencies
# Install: pip install -r requirements.txt

opencv-python>=4.8.0
mediapipe>=0.10.9
onnxruntime>=1.16.0
faiss-cpu>=1.7.4
streamlit>=1.30.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
filterpy>=1.4.5
scikit-learn>=1.3.0
pyttsx3